[TS-333] Refactor: 로딩스피너 파일 위치 변경
